changed: put bookmark videodatabase methods in separate file
